test(next-drupal): fix bug in jsonRpc test util and rename utils files
