+++ packages/next-drupal/tests/utils/mocks/index.ts
@@ -1,3 +1,3 @@
 export * from "./data"
-export * from "./fetch"
-export * from "./logger"
+export * from "./spyOnFetch"
+export * from "./mockLogger"
